% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt.R
\name{cpt.np}
\alias{cpt.np}
\title{Identifying Changes using a Nonparametric Cost Function}
\usage{
cpt.np(data, penalty = "MBIC", pen.value = 0, method = "PELT",
  test.stat = "empirical_distribution", class = TRUE, minseglen = 1,
  test.param = 10)
}
\arguments{
\item{data}{A vector, ts object or matrix containing the data within which you wish to find a changepoint.  If the data is a matrix, each row is considered as a separate dataset.}

\item{penalty}{Choice of "None", "SIC", "BIC", "MBIC", AIC", "Hannan-Quinn", "Manual" and "CROPS" penalties.  If Manual is specified, the manual penalty is contained in the pen.value parameter. If CROPS is specified, the penalty range is contained in the pen.value parameter; note this is a vector of length 2 which contains the minimum and maximum penalty value.  Note CROPS can only be used if the method is "PELT". The predefined penalties listed DO count the changepoint as a parameter, postfix a 0 e.g."SIC0" to NOT count the changepoint as a parameter.}

\item{pen.value}{The value of the penalty when using the Manual penalty option.  A vector of length 2 (min,max) if using the CROPS penalty.}

\item{method}{There are two methods available "PELT" and "FPOP".}

\item{test.stat}{The assumed test statistic/distribution of the data. If using the "PELT" method, currently only "empirical_distribution" is available. For the "FPOP" method, "L1", "Huber", and "Outlier" are available.}

\item{class}{Logical. If TRUE then an object of class cpt is returned.}

\item{minseglen}{Positive integer giving the minimum segment length (number of observations between changes), default is the minimum allowed by theory.}

\item{test.param}{Additional values required by methods and/or distributions. .For test.stat = "empirical_distribution", this is the number of quantiles to calculate; For method == "FPOP", this is the lthreshold.}
}
\value{
If \code{class=TRUE} then an object of S4 class "cpt" is returned.  The slot \code{cpts} contains the changepoints that are returned.  For \code{class=FALSE} the structure is as follows.

If data is a vector (single dataset) then a vector/list is returned depending on the value of method.  If data is a matrix (multiple datasets) then a list is returned where each element in the list is either a vector or list depending on the value of method.

If method is PELT then a vector is returned containing the changepoint locations for the penalty supplied. If the penalty is CROPS then a list is returned with the elements:

\item{cpt.out}{A data frame containing the value of the penalty value where the number of segmentations chages, the number of segmentations and the value of the cost at that penalty value.}
\item{changepoints}{The optimal changepoints for the different penalty values startings with the lowest penalty value.}
}
\description{
Calculates the optimal positioning and number of changepoints for data given a user specified cost function and penalty.
}
\details{
This function is used to find multiple changes in a data set using either of two changepoint algorithms. The two choices are:PELT with a nonparametric cost function based on the empirical distribution; and "FPOP" with a nonparametric cost function based on the L1, Huber or Outlier loss.  A changepoint is denoted as the first observation of the new segment.
}
\references{
PELT with an Empirical Distribution cost function: Haynes K, Fearnhead P, Eckley I A (2016) A computationally efficient nonparametric approach for changepoint detection, Statistics and Computed (accepted)

PELT Algorithm: Killick R, Fearnhead P, Eckley I A (2012) Optimal detection of changepoints with a linear computational cost, \emph{JASA 107(500), 1590-1598}

CROPS: Haynes K, Eckley I A, Fearnhead P (2015) Computationally Efficient Changepoint Detection for a Range of Penalties, \emph{JCGS, To Appear}

FPOP with L1, Huber and Outlier Losses: Paul Fearnhead, Guillem Rigaill (2016) Changepoint Detection in the Presence of Outliers

FPOP: Robert Maidstone, Toby Hocking, Guillem Rigaill, Paul Fearnhead (2014) On Optimal Multiple Changepoint Algorithms for Large Data
}
\seealso{
PELT in parametric settings: \code{\link[changepoint]{cpt.mean}} for changes in the mean, \code{\link[changepoint]{cpt.var}} for changes in the variance and \code{\link[changepoint]{cpt.meanvar}} for changes in the mean and variance.
}
\author{
Kaylea Haynes
}
