% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeartRate.R
\docType{data}
\name{HeartRate}
\alias{HeartRate}
\title{Recorded heart rate during a run}
\format{A vector of recorded heart rates at points over time with 1160 data points.}
\usage{
HeartRate
}
\description{
A dataset containing heart-rate recorded duing a run.
}
\author{
Kaylea Haynes
}
\references{
For details see: Haynes K, Fearnhead P, Eckley I A (2016) A computationally efficient nonparametric approach for changepoint detection, Statistics and Computed (accepted)
}
\keyword{datasets}

