% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HeartRate.R
\docType{data}
\name{HeartRate}
\alias{HeartRate}
\title{Recorded heart rate during a run}
\format{
A vector of recorded heart rates at points over time with 1160 data points.
}
\usage{
data(HeartRate)
}
\description{
A dataset containing heart-rate recorded during a run.
}
\references{
\insertRef{Haynes2017}{changepoint.np}
}
\author{
Kaylea Haynes
}
\keyword{datasets}
